jpa upadte

+++ jungocrawling/src/main/java/com/jungo/jungocrawling/Account/ItemRepository.java
@@ -28,7 +28,8 @@ public interface ItemRepository extends JpaRepository<Item, Long> {
 
     List<Item> findByTitleContainsOrderByIdDesc(String title, Pageable pageable);
 
-    long countByTitleContains(String title);
+    int countByTitleContains(String title);
+
 
     @Query(nativeQuery = true, value = "select * from item where title LIKE CONCAT('%', :keyword, '%') ORDER BY price")
     List<Item> findByASC(String keyword);
